webapi/tct-webaudio-w3c-tests/webaudio/automation-support.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.
Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:
* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.
THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>Web Audio: dynamicscompressornode_reduction_default_value</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chenglei Ren" href="mailto:chenglei.ren@intel.com">
<link rel="help" href="https://webaudio.github.io/web-audio-api">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/webaudio.js"></script>
<div id="log"></div>
<script>
    test(function () {
        var context = new AudioContext();
        var pannerNode = context.createPanner();
        assert_true("positionX" in pannerNode, "Check if PannerNode.positionX attribute exists");
        assert_true(pannerNode.positionX instanceof AudioParam, "Check if PannerNode.positionX attribute is AudioParam type");
        assert_equals(pannerNode.positionX.defaultValue, 0, "Check if PannerNode.positionX default value is 0");
        assert_true("positionY" in pannerNode, "Check if PannerNode.positionY attribute exists");
        assert_true(pannerNode.positionY instanceof AudioParam, "Check if PannerNode.positionY attribute is AudioParam type");
        assert_equals(pannerNode.positionY.defaultValue, 0, "Check if PannerNode.positionY default value is 0");
        assert_true("positionZ" in pannerNode, "Check if PannerNode.positionZ attribute exists");
        assert_true(pannerNode.positionZ instanceof AudioParam, "Check if PannerNode.positionZ attribute is AudioParam type");
        assert_equals(pannerNode.positionZ.defaultValue, 0, "Check if PannerNode.positionZ default value is 0");
    }, "Check if PannerNode.position attribute is supported");
    test(function () {
        var context = new AudioContext();
        var pannerNode = context.createPanner();
        assert_true("orientationX" in pannerNode, "Check if PannerNode.orientationX attribute exists");
        assert_true(pannerNode.orientationX instanceof AudioParam, "Check if PannerNode.orientationX attribute is AudioParam type");
        assert_equals(pannerNode.orientationX.defaultValue, 1, "Check if PannerNode.orientationX default value is 0");
        assert_true("orientationY" in pannerNode, "Check if PannerNode.orientationY attribute exists");
        assert_true(pannerNode.orientationY instanceof AudioParam, "Check if PannerNode.orientationY attribute is AudioParam type");
        assert_equals(pannerNode.orientationY.defaultValue, 0, "Check if PannerNode.orientationY default value is 0");
        assert_true("orientationZ" in pannerNode, "Check if PannerNode.orientationZ attribute exists");
        assert_true(pannerNode.orientationZ instanceof AudioParam, "Check if PannerNode.orientationZ attribute is AudioParam type");
        assert_equals(pannerNode.orientationZ.defaultValue, 0, "Check if PannerNode.orientationZ default value is 0");
    }, "Check if PannerNode.orientation attribute is supported");
    test(function () {
        var context = new AudioContext();
        var listener = context.listener;
        assert_true("positionX" in listener, "Check if listener.positionX attribute exists");
        assert_true(listener.positionX instanceof AudioParam, "Check if listener.positionX attribute is AudioParam type");
        assert_equals(listener.positionX.defaultValue, 0, "Check if listener.positionX default value is 0");
        assert_true("positionY" in listener, "Check if listener.positionY attribute exists");
        assert_true(listener.positionY instanceof AudioParam, "Check if listener.positionY attribute is AudioParam type");
        assert_equals(listener.positionY.defaultValue, 0, "Check if listener.positionY default value is 0");
        assert_true("positionZ" in listener, "Check if listener.positionZ attribute exists");
        assert_true(listener.positionZ instanceof AudioParam, "Check if listener.positionZ attribute is AudioParam type");
        assert_equals(listener.positionZ.defaultValue, 0, "Check if listener.positionZ default value is 0");
    }, "Check if listener.position attribute is supported");
    test(function () {
        var context = new AudioContext();
        var listener = context.listener;
        assert_true("forwardX" in listener, "Check if listener.forwardX attribute exists");
        assert_true(listener.forwardX instanceof AudioParam, "Check if listener.forwardX attribute is AudioParam type");
        assert_equals(listener.forwardX.defaultValue, 0, "Check if listener.forwardX default value is 0");
        assert_true("forwardY" in listener, "Check if listener.forwardY attribute exists");
        assert_true(listener.forwardY instanceof AudioParam, "Check if listener.forwardY attribute is AudioParam type");
        assert_equals(listener.forwardY.defaultValue, 0, "Check if listener.forwardY default value is 0");
        assert_true("forwardZ" in listener, "Check if listener.forwardZ attribute exists");
        assert_true(listener.forwardZ instanceof AudioParam, "Check if listener.forwardZ attribute is AudioParam type");
        assert_equals(listener.forwardZ.defaultValue, -1, "Check if listener.forwardZ default value is 0");
    }, "Check if listener.forward attribute is supported");
    test(function () {
        var context = new AudioContext();
        var listener = context.listener;
        assert_true("upX" in listener, "Check if listener.upX attribute exists");
        assert_true(listener.upX instanceof AudioParam, "Check if listener.upX attribute is AudioParam type");
        assert_equals(listener.upX.defaultValue, 0, "Check if listener.upX default value is 0");
        assert_true("upY" in listener, "Check if listener.upY attribute exists");
        assert_true(listener.upY instanceof AudioParam, "Check if listener.upY attribute is AudioParam type");
        assert_equals(listener.upY.defaultValue, 1, "Check if listener.upY default value is 0");
        assert_true("upZ" in listener, "Check if listener.upZ attribute exists");
        assert_true(listener.upZ instanceof AudioParam, "Check if listener.upZ attribute is AudioParam type");
        assert_equals(listener.upZ.defaultValue, 0, "Check if listener.upZ default value is 0");
     }, "Check if listener.up attribute is supported");
</script>
